add buttons for Leap / Tumbleweed portals
Once a visitor has read the mouseover text for Leap or Tumbleweed, it is
likely they will want to learn more about Leap or Tumbleweed, so make it
easy for them to find the relevant Portal page.

Also clarify that the current stable release is 13.2, since 42.1 has not
yet been released.