Update index.html
Imprint/Impressum is a legal requirement (German TMG) and we need to make sure there is a link to the relevant legal information directly from the homepage.

Note that only the german wiki has the information. There is an english Terms_of_site page but it doesn't contain the same information as the german one. It may make sense to update the english wiki too.

This is just a suggestion, but the link needs to be on the front page.