changing the counter to microos link
Having two Leap options on the landing page is redundant. Issue about not having microos on the landing page is solved and will help people find it. The previous setup would not give the option for people to know about it because the links for leap and tw automatically go to get.o.o /leap /tw.